Migrate path examples to new structure
