+++ examples/path/introduction/Cargo.toml
@@ -1,7 +1,9 @@
 [package]
-name = "basic_path_extractor"
-version = "0.1.0"
+name = "gotham_examples_path_introduction"
+version = "0.0.0"
+description = "An introduction to extracting request path segments, in a type safe way, with the Gotham web framework"
 authors = ["Nicolas Pochet <npochet@gmail.com>"]
+publish = false
 
 [dependencies]
 gotham = { path = "../../../gotham" }
